Adds helper method for test (test_action test_intent test_attack etc) Adds passing test replace i32 -> i64

+++ rust/game-manager/game_manager/src/game_manager_mod.rs
@@ -15,7 +15,7 @@ pub struct EntityId(pub i64);
 #[derive(Debug, PartialEq, PartialOrd, Eq, Ord, Hash, Clone, Copy, ToVariant)]
 pub enum TeamId {
     /// a unique identifier for each team
-    Team(i32),
+    Team(i64),
     /// the entity has no team and can attack anyone
     Loner,
 }
